moved the getTotalAmountWithInterestAndFees to claimDeatilsService and included unit test
