+++ src/main/modules/claimDetailsService.ts
@@ -1,0 +1,6 @@
+import {Claim} from '../common/models/claim';
+import {convertToPoundsFilter} from '../common/utils/currencyFormat';
+
+export const getTotalAmountWithInterestAndFees = (claim: Claim) => {
+  return claim.totalClaimAmount + claim.totalInterest + convertToPoundsFilter(claim.claimFee.calculatedAmountInPence);
+};
