+++ src/main/views/macro/amount-breakdown.njk
@@ -41,8 +41,8 @@
       {% endif %}
       </tr>
       <tr class="govuk-table__row">
-        <td class="govuk-table__cell"><strong>{{ t('PAGES.CLAIM_DETAILS.CLAIM_TOTAL') }}</strong></td>
-        <td class="govuk-table__cell govuk-!-text-align-right"><strong>{{ totalAmount | currencyFormat(totalAmount) }}</strong></td>
+        <td class="govuk-table__cell govuk-!-font-weight-bold">{{ t('PAGES.CLAIM_DETAILS.CLAIM_TOTAL') }}</td>
+        <td class="govuk-table__cell govuk-!-font-weight-bold govuk-!-text-align-right">{{ totalAmount | currencyFormat(totalAmount) }}</td>
       </tr>
   </tbody>
 </table>
